+++ src/main/java/com/example/tickets/repository/TicketRepository.java
@@ -2,13 +2,10 @@ package com.example.tickets.repository;
 
 import com.example.tickets.data.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
 import java.util.Optional;
 
 public interface TicketRepository extends JpaRepository<Ticket, Long> {
 
-    Optional<Ticket> findByNumber(int number);
+    Optional<Ticket> findByNumber(Long number);
     Ticket findTopByOrderByNumberDesc();
 }
